fix(PR-2116): keep lookup join inbound guard in broker
